+++ scripts/vllm-script/.gitignore
@@ -1,0 +1,2 @@
+data
+testing*
